Move PolyFuncType into poly_func.rs
